refactor: move provider dashboard to modules
refs #1272
