+++ apps/api/src/routes/internal/index.ts
@@ -2,8 +2,7 @@ import gpu from "../v1/gpu";
 import gpuModels from "../v1/gpuModels";
 import gpuPrices from "../v1/gpuPrices";
 import leasesDuration from "../v1/leasesDuration";
-import providerDashboard from "../v1/providerDashboard";
 import providerVersions from "../v1/providerVersions";
 import financial from "./financial";
 
-export default [providerVersions, gpu, leasesDuration, gpuModels, gpuPrices, providerDashboard, financial];
+export default [providerVersions, gpu, leasesDuration, gpuModels, gpuPrices, financial];
